cancel source task directly after timeout
Why:
The source task may still suceed after a timeout if the backup task
does not complete immediatily,
e.g. et=execution time, source-et < timeout + backup-et

This is contrary to the documentation and the intended behaviour.

Side-effects:
This is a breaking change, because as stated above code will return
in some cases the result of a different task.