Add Alternative instance for Observable
There are instances for Applicative and MonoidK, but not for
Alternative. Observable already implements all the necessary abstract
methods and follows the Alternative laws.

This adds Alternative to Observable.CatsInstances so that Alternative
methods like `guard` can be used. It's also more convenient as an
implicit dependency than `MonoidK with Applicative`.